remove some casts to short
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1825749 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xssf/streaming/SXSSFCell.java
@@ -613,7 +613,7 @@ public class SXSSFCell implements Cell {
     {
         if(_style == null){
             SXSSFWorkbook wb = (SXSSFWorkbook)getRow().getSheet().getWorkbook();
-            return wb.getCellStyleAt((short)0);
+            return wb.getCellStyleAt(0);
         } else {
             return _style;
         }
